[♻️refactor] : 스타일링, 코드, 폴더 구조 수정
