+++ src/components/MyPosts/style.ts
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+
+export const PostWrapper = styled.div`
+  margin-top: 3rem;
+`;
+
+export const EmptyPostTitle = styled.div`
+  margin-left: 1rem;
+  color: #777777;
+`;
