+++ src/components/UserInfo/style.ts
@@ -1,0 +1,17 @@
+import styled from '@emotion/styled';
+
+export const UserInfoSection = styled.div`
+  margin-left: 2.2rem;
+  margin-top: 3rem;
+`;
+
+export const UserName = styled.div`
+  font-size: 1.8rem;
+  margin-bottom: 0.4rem;
+  font-weight: bold;
+`;
+
+export const UserId = styled.div`
+  font-size: 1.3rem;
+  color: #9796a1;
+`;
